Addresses issue #334
